stock/requirements.txt
# Core requirements
praw>=7.7.1
streamlit>=1.35.0
requests>=2.28.1
tweepy>=4.10.0

# Machine Learning and NLP
torch>=2.3.1
transformers>=4.12.3
spacy>=3.1.2
vaderSentiment>=3.3.2


# Additional utilities
scikit-learn>=1.0.2

# Spacy language model
spacy==3.7.5
pipreqs
psycopg2-binary
yfinance
# Installation via Conda Forge
# Note: faiss-cpu is not available as 'faiss' on Conda Forge
# conda install -c conda-forge faiss-cpu

# then run
# python -m spacy download en_core_web_sm